Move TestSpecificImageNameSubstitutor back to original package and remove duplicate use of default substitutor
