+++ docs/examples/junit4/generic/src/test/resources/testcontainers.properties
@@ -1,1 +1,1 @@
-image.substitutor=org.testcontainers.utility.TestSpecificImageNameSubstitutor
+image.substitutor=generic.support.TestSpecificImageNameSubstitutor
